python/tests/requirements.txt
pytest>=7.4.0
pytest-cov>=4.0.0
pytest-xdist>=3.3.0
pytest-html>=3.1.0
pytest-benchmark>=4.0.0
pytest-mock>=3.11.0
pyyaml>=6.0
coverage>=7.0.0
psutil>=5.9.0
